Fix duration type in TemplateHistoryResponse
- Change duration from int to float
- Matches database Numeric(8, 2) type
- Fixes Pydantic validation error for decimal durations

+++ app/api/templates.py
@@ -20,7 +20,7 @@ class TemplateHistoryResponse(BaseModel):
     id: str
     title: str
     hotel_name: str
-    duration: int
+    duration: float
     script: str
     video_link: Optional[str] = None
     viewed_at: Optional[str] = None
